refactor: remove redundant type casts and make casts from int to long explicit

+++ src/main/java/dev/galacticraft/mod/client/render/dimension/VenusWeatherRenderer.java
@@ -111,7 +111,7 @@ public class VenusWeatherRenderer implements DimensionRenderingRegistry.WeatherR
                     }
 
                     if (y != ymax) {
-                        RandomSource random = RandomSource.create((long) (x * x * 3121 + x * 45238971 ^ z * z * 418711 + z * 13761));
+                        RandomSource random = RandomSource.create(x * x * 3121L + x * 45238971L ^ z * z * 418711L + z * 13761L);
 
                         if (drawFlag != 0) {
                             if (drawFlag >= 0 && buffer != null) {
